backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiohttp==3.9.5
python-dotenv==1.0.1
sqlalchemy==2.0.23
pydantic==2.5.0